Split out VSO exceptions into their own file
